% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.tree.R
\name{copy.gates}
\alias{copy.gates}
\title{copy the gates to the existing parents}
\usage{
copy.gates(gs, parents, gate.paths)
}
\value{
the new sub tree paths
}
\description{
copy the gates to the existing parents
}

