% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.tree.R
\name{find.gates}
\alias{find.gates}
\title{find the gate paths that terminates with the given marker}
\usage{
find.gates(marker, all.nodes)
}
\arguments{
\item{marker}{marker name that is associated with the target gate}

\item{all.nodes}{the candidate gating paths to be searched from}
}
\description{
find the gate paths that terminates with the given marker
}
\examples{
\dontrun{
find.gates("CD8", getNodes(gs))
#expect to return something like "/LYM/singlets/CD8+" and "/LYM/singlets/CD8-"
}

}
