% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get.openCyto.exhaustive}
\alias{get.openCyto.exhaustive}
\title{get the enviornment that stores the marker selection results}
\usage{
get.openCyto.exhaustive(gs, node)
}
\arguments{
\item{gs}{GatingSet}

\item{node}{the unique path to the gated population node}
}
\description{
get the enviornment that stores the marker selection results
}

