% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getMetrics}
\alias{getMetrics}
\title{get the metrics table computed by the marker selection method}
\usage{
getMetrics(...)
}
\arguments{
\item{...}{arguments
gs GatingSet
node the unique path to the gated population node}
}
\description{
get the metrics table computed by the marker selection method
}
