% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.tree.R
\name{interpolate.tree}
\alias{interpolate.tree}
\title{add gates to the gating tree so that it is balanced complete tree}
\usage{
interpolate.tree(gs)
}
\arguments{
\item{gs}{GatingSet}
}
\description{
It is done by copying the gates from the nearest the neighouring branches to the path where the respective
marker is missing from the existing path. So that the gating tree is more balanced.
}
