% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isGated}
\alias{isGated}
\title{check if the marker has already been gated to avoid gating on the same marker repeately on the same path}
\usage{
isGated(marker, gate.path)
}
\arguments{
\item{marker}{marker name to check}

\item{gate.path}{the gating path that contains the seriers of markers}
}
\description{
check if the marker has already been gated to avoid gating on the same marker repeately on the same path
}
\examples{
 isGated("CD3 APC", "/boundary/nonDebris/lymph/CD19-/CD3+")
 isGated("CD38", "/boundary/nonDebris/lymph/CD19-/CD3+")
}

