% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{plot.flowClust}
\alias{plot.flowClust}
\title{plot the flowClust fitting results for a particular gated node}
\usage{
plot.flowClust(...)
}
\arguments{
\item{...}{arguments
gs GatingSet
node the unique path to the gated population node}
}
\description{
plot the flowClust fitting results for a particular gated node
}
